xenclient-oe: networkmanager: fix issues related to devices managed by networkmanager
OXT-1281

NetworkManager.conf: The new configuration filters the devices to be manager by networkmanager.
		     We don't want all the devices to be managed (for ex: brshared, brany, etc.., ).

auto-dhcp-for-eth0: Since eth0 is added as a bridge slave for brbridged dhcp is disabled by default on eth0, this patch enables it.

fix-device-type-property-exposure: This fixes the device type property exposed to clients like nm-applet, nmcli, etc..,

Signed-off-by: Mahantesh Salimath <salimathm@ainfosec.com>